pyproject.toml
[tool.poetry]
name = "PoetryEnvManagingPackages"
version = "0.1.0"
description = ""
authors = ["Natalia Stanko <anithaly@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.23.0"
python-box = "^4.2.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
